chore: rename nested composable file, allow nested items to live outside nested provide
