+++ packages/vuetify/src/components/VNested/VNested.tsx
@@ -1,4 +1,4 @@
-import { useNested, useNestedGroup, useNestedItem } from '@/composables/nested'
+import { useNested, useNestedGroup, useNestedItem } from '@/composables/nested/nested'
 import { defineComponent } from '@/util'
 import { toRef } from 'vue'
